+++ src/optimumai/__init__.py
@@ -224,7 +224,7 @@ from optimumai.visualization.plotstudio import (
 )
 from optimumai.world_models.jepa import JEPA
 
-__version__ = "1.6.1"
+__version__ = "1.6.2"
 
 __all__ = [
     "COURSE",